Generalise readFromAirbyte
